Se realiza la lógica de recate de medicaciones del usuario

+++ ObesityIOS/ObesityIOS/entities/medication/Medication.swift
@@ -10,23 +10,23 @@ import Foundation
 
 class Medication {
     
-    var timestamp:Int64 = 0
+    var timestamp:Double = 0
     var iduser:String = ""
     var medicine:String = ""
     var observations:String = ""
-    var begin_timestamp:Int64 = 0
-    var end_timestamp:Int64 = 0
+    var begin_timestamp:Double = 0
+    var end_timestamp:Double = 0
     var method:String = ""
     var days:String = ""
     var dosage:String = ""
     var createdBy:String = ""
-    var userViewTimestamp:Int64 = 0
+    var userViewTimestamp:Double = 0
     var sideEffects:[MedicationSideEffect] = []
     
     init() {
     }
     
-    init(timestamp:Int64, iduser:String, medicine:String, observations:String, begin_timestamp:Int64, end_timestamp:Int64, method:String, days:String, dosage:String, createdBy:String, userViewTimestamp:Int64, sideEffects:[MedicationSideEffect]) {
+    init(timestamp:Double, iduser:String, medicine:String, observations:String, begin_timestamp:Double, end_timestamp:Double, method:String, days:String, dosage:String, createdBy:String, userViewTimestamp:Double, sideEffects:[MedicationSideEffect]) {
         
         self.timestamp = timestamp
         self.iduser = iduser
